+++ ci/check_login.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -ex
+
+cd $(dirname $0)
+
+vagrant ssh -- sudo /vagrant/remote/check_login.sh "$@"