+++ jenkins.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -ex
+
+./build.sh
+
+summon -f secrets.ci.yml ./test.sh "$@"