+++ recipes/_configure_logshipper.rb
@@ -1,8 +1,9 @@
 service 'logshipper' do
-  provider Chef::Provider::Service::Upstart
   action :nothing
-
-  only_if { node['conjur']['service_provider'] == "upstart" }
+  if node['platform'] == 'amazon' ||
+      (node['platform_family'] == 'rhel' && ConjurDetect.platform_version?(node, '~>6.0'))
+    provider Chef::Provider::Service::Upstart
+  end
 end
 
 file "/etc/conjur.identity" do
@@ -14,6 +15,6 @@ file "/etc/conjur.identity" do
 machine #{conjur_appliance_url}/authn
     login host/#{conjur_host_id}
     password #{conjur_host_api_key}
-  """
-  notifies :restart, 'service[logshipper]', :delayed if node['conjur']['service_provider'] == "upstart"
+"""
+  notifies(:restart, 'service[logshipper]', :delayed)
 end