+++ recipes/_install_logshipper_upstart.rb
@@ -1,7 +1,3 @@
-service 'logshipper' do
-  provider Chef::Provider::Service::Upstart
-end
-
 command_line = "/usr/sbin/logshipper -n #{logshipper_fifo_path} >> /var/log/logshipper.log 2>&1"
 
 # differentiate here the platform-specific parts