+++ recipes/_install_ssh_rhel.rb
@@ -8,7 +8,7 @@ package_list = %W(
   oddjob
 )
 
-if Chef::VersionConstraint.new('~> 7.0').include?(node['platform_version'])
+if ConjurDetect.platform_version?(node, '~>7.0')
   # this is needed for mkhomedir to work on rhel 7 and variants
   package_list << 'oddjob-mkhomedir'
   
@@ -39,7 +39,7 @@ end
 # CentOS and RHEL actually have '6' or '7' in the variable.
 releasever = {
   'amazon' => '7'
-}[node.platform] || '$releasever'
+}[node['platform']] || '$releasever'
 
 yum_repository 'conjur' do
   description 'Conjur Inc.'