+++ recipes/identity.rb
@@ -25,5 +25,5 @@ file "/etc/conjur.identity" do
 machine #{conjur_appliance_url}/authn
     login host/#{conjur_host_id node}
     password #{conjur_host_api_key node}
-  """
+"""
 end.run_action(:create)